First Commit
These files were built in MIMO, copy pasta to KODEX. Saved to FILES.
THe REPO made in GIthub and cloned on Working Copy. This is the
initial commit of 5 files/ .. The titles of the files have changed. To
add spr1ng_tech_% ... the html file was originally called index.html.
All code imports and exports will need to be updated to reflect so
they are not broken.